+++ src/read_config_file.F90
@@ -52,7 +52,7 @@ subroutine read_config_file(Ctrl, channel_ids_instr, channel_sw_flag, &
    integer, allocatable, dimension(:), intent(inout) :: channel_ids_instr
    integer, allocatable, dimension(:), intent(inout) :: channel_sw_flag
    integer, allocatable, dimension(:), intent(inout) :: channel_lw_flag
-   integer, allocatable, dimension(:), intent(inout) :: channel_wvl
+   real,    allocatable, dimension(:), intent(inout) :: channel_wvl
    type(global_attributes_s),          intent(inout) :: global_atts
    type(source_attributes_s),          intent(inout) :: source_atts
    logical,                            intent(in)    :: verbose
